Remove unused fields from invoice billing address
