+++ app/models/invoice_creator.rb
@@ -50,7 +50,7 @@ class InvoiceCreator
   end
 
   def assign_billing_address
-    billing_fields = %w[vat_code legal_entity street city state postal_code alpha_two_country_code]
+    billing_fields = %w[vat_code street city postal_code alpha_two_country_code]
     invoice.recipient = invoice.billing_profile.name
 
     invoice.billing_profile.attributes.keys.each do |attribute|
